refactor(core): split layout formatters, axis train, segments, and types
Extract layout domain/formatters/ticks; band vs continuous axis training;
segment emitters; frame binding/LayerFrame types; geom/stat contracts;
candidate auto-mode; scene panel assembly; panel placements; sequential
legend format; and candidate field-channel helpers. Facades keep import
paths stable. Add axisTicks characterization.
