+++ packages/core/src/pipeline/bind-layer-type-contracts.ts
@@ -1,11 +1,12 @@
 /**
  * Geom/stat channel type mismatch and computed-y contracts for bindLayer.
  */
-import type { BarParams, LayerSpec } from "@ggsvelte/spec";
+import type { LayerSpec } from "@ggsvelte/spec";
 
 import type { ColumnTable } from "../table.js";
 
-import { PipelineError } from "./types.js";
+import { validateComputedYAndBinContracts } from "./bind-layer-type-contracts-computed.js";
+import { validateGeomChannelTypeContracts } from "./bind-layer-type-contracts-geom.js";
 
 export function validateGeomStatContracts(input: {
   layer: LayerSpec;
@@ -14,77 +15,6 @@ export function validateGeomStatContracts(input: {
   xField: string | null;
   yField: string | null;
 }): void {
-  const { layer, index, table, xField, yField } = input;
-  const geom = layer.geom;
-  const stat = layer.stat ?? "identity";
-  const params = layer.params ?? {};
-
-  if (geom === "bar" && yField !== null) {
-    throw new PipelineError(
-      "computed-y-mapped",
-      `/layers/${index}/aes/y`,
-      `The bar geom computes y with the ${stat} stat, so aes.y must not map data. Use geom "col" for pre-computed heights.`,
-    );
-  }
-  if (geom === "density" && yField !== null) {
-    throw new PipelineError(
-      "computed-y-mapped",
-      `/layers/${index}/aes/y`,
-      "The density geom computes y with the density stat, so aes.y must not map data. Map only x.",
-    );
-  }
-  if (stat === "bin") {
-    const p = params as BarParams;
-    if (p.center !== undefined && p.boundary !== undefined) {
-      throw new PipelineError(
-        "bin-center-and-boundary",
-        `/layers/${index}/params`,
-        "The bin stat accepts params.center OR params.boundary (both align the bin grid), never both.",
-      );
-    }
-    if (xField !== null && table.fieldType(xField) === "nominal") {
-      throw new PipelineError(
-        "channel-type-mismatch",
-        `/layers/${index}/aes/x`,
-        `The bin stat needs a continuous x, but field "${xField}" is nominal. Use geom "bar" (the count stat) to count categories instead.`,
-      );
-    }
-  }
-  if (geom === "density" && xField !== null && table.fieldType(xField) === "nominal") {
-    throw new PipelineError(
-      "channel-type-mismatch",
-      `/layers/${index}/aes/x`,
-      `The density stat needs a continuous x, but field "${xField}" is nominal. Use geom "bar" (the count stat) to count categories instead.`,
-    );
-  }
-  if (geom === "smooth") {
-    for (const [channel, field] of [
-      ["x", xField],
-      ["y", yField],
-    ] as const) {
-      if (field !== null && table.fieldType(field) === "nominal") {
-        throw new PipelineError(
-          "channel-type-mismatch",
-          `/layers/${index}/aes/${channel}`,
-          `The smooth stat needs quantitative x and y, but field "${field}" (${channel}) is nominal.`,
-        );
-      }
-    }
-  }
-  if (geom === "boxplot") {
-    if (xField !== null && table.fieldType(xField) !== "nominal") {
-      throw new PipelineError(
-        "channel-type-mismatch",
-        `/layers/${index}/aes/x`,
-        `The boxplot geom needs a DISCRETE x this milestone, but field "${xField}" is ${table.fieldType(xField)}. Map x to a categorical field.`,
-      );
-    }
-    if (yField !== null && table.fieldType(yField) === "nominal") {
-      throw new PipelineError(
-        "channel-type-mismatch",
-        `/layers/${index}/aes/y`,
-        `The boxplot stat needs a quantitative y, but field "${yField}" is nominal.`,
-      );
-    }
-  }
+  validateComputedYAndBinContracts(input);
+  validateGeomChannelTypeContracts(input);
 }
