+++ packages/core/src/pipeline/frame-candidates-auto-mode.ts
@@ -1,0 +1,36 @@
+/**
+ * Default inspection mode for a bound layer mark.
+ */
+import type { ResolvedCandidateInspectMode } from "../candidate-store.js";
+
+import { interceptList } from "./frame-helpers.js";
+import type { LayerBinding } from "./types.js";
+
+export function candidateAutoMode(
+  binding: LayerBinding,
+  primitiveIndex: number,
+): ResolvedCandidateInspectMode {
+  switch (binding.layer.geom) {
+    case "point":
+    case "text":
+      return "xy";
+    case "col":
+    case "bar":
+      return "exact";
+    case "line":
+    case "area":
+    case "density":
+    case "smooth":
+    case "errorbar":
+    case "boxplot":
+      return "x";
+    case "rule": {
+      if (binding.ruleForm === "vertical") return "x";
+      if (binding.ruleForm === "horizontal") return "y";
+      const params = (binding.layer.params ?? {}) as { xintercept?: unknown };
+      return primitiveIndex < interceptList(params.xintercept).length ? "x" : "y";
+    }
+    default:
+      return "xy";
+  }
+}
