+++ packages/core/src/pipeline/types-no-row.ts
@@ -1,0 +1,4 @@
+/**
+ * Sentinel source-row index for synthesized geometry (stat-only marks).
+ */
+export const NO_ROW = 0xffffffff;
